'Refactored by Sourcery'

+++ GabiBraunRobot/modules/cleaner.py
@@ -10,11 +10,7 @@ from telegram import ParseMode, Update
 from telegram.ext import (CallbackContext, CommandHandler, Filters,
                           MessageHandler, run_async)
 
-if ALLOW_EXCL:
-    CMD_STARTERS = ('/', '!')
-else:
-    CMD_STARTERS = ('/')
-
+CMD_STARTERS = ('/', '!') if ALLOW_EXCL else '/'
 BLUE_TEXT_CLEAN_GROUP = 13
 CommandHandlerList = (CommandHandler, CustomCommandHandler,
                       DisableAbleCommandHandler)
@@ -37,22 +33,21 @@ def clean_blue_text_must_click(update: Update, context: CallbackContext):
     bot = context.bot
     chat = update.effective_chat
     message = update.effective_message
-    if chat.get_member(bot.id).can_delete_messages:
-        if sql.is_enabled(chat.id):
-            fst_word = message.text.strip().split(None, 1)[0]
+    if chat.get_member(bot.id).can_delete_messages and sql.is_enabled(chat.id):
+        fst_word = message.text.strip().split(None, 1)[0]
 
-            if len(fst_word) > 1 and any(
-                    fst_word.startswith(start) for start in CMD_STARTERS):
+        if len(fst_word) > 1 and any(
+                fst_word.startswith(start) for start in CMD_STARTERS):
 
-                command = fst_word[1:].split('@')
-                chat = update.effective_chat
+            command = fst_word[1:].split('@')
+            chat = update.effective_chat
 
-                ignored = sql.is_command_ignored(chat.id, command[0])
-                if ignored:
-                    return
+            ignored = sql.is_command_ignored(chat.id, command[0])
+            if ignored:
+                return
 
-                if command[0] not in command_list:
-                    message.delete()
+            if command[0] not in command_list:
+                message.delete()
 
 
 @run_async
@@ -67,14 +62,14 @@ def set_blue_text_must_click(update: Update, context: CallbackContext):
         val = args[0].lower()
         if val in ('off', 'no'):
             sql.set_cleanbt(chat.id, False)
-            reply = "Bluetext cleaning has been disabled for <b>{}</b>".format(
-                html.escape(chat.title))
+            reply = f"Bluetext cleaning has been disabled for <b>{html.escape(chat.title)}</b>"
+
             message.reply_text(reply, parse_mode=ParseMode.HTML)
 
         elif val in ('yes', 'on'):
             sql.set_cleanbt(chat.id, True)
-            reply = "Bluetext cleaning has been enabled for <b>{}</b>".format(
-                html.escape(chat.title))
+            reply = f"Bluetext cleaning has been enabled for <b>{html.escape(chat.title)}</b>"
+
             message.reply_text(reply, parse_mode=ParseMode.HTML)
 
         else:
@@ -82,12 +77,9 @@ def set_blue_text_must_click(update: Update, context: CallbackContext):
             message.reply_text(reply)
     else:
         clean_status = sql.is_enabled(chat.id)
-        if clean_status:
-            clean_status = "Enabled"
-        else:
-            clean_status = "Disabled"
-        reply = "Bluetext cleaning for <b>{}</b> : <b>{}</b>".format(
-            html.escape(chat.title), clean_status)
+        clean_status = "Enabled" if clean_status else "Disabled"
+        reply = f"Bluetext cleaning for <b>{html.escape(chat.title)}</b> : <b>{clean_status}</b>"
+
         message.reply_text(reply, parse_mode=ParseMode.HTML)
 
 
@@ -101,8 +93,7 @@ def add_bluetext_ignore(update: Update, context: CallbackContext):
         val = args[0].lower()
         added = sql.chat_ignore_command(chat.id, val)
         if added:
-            reply = "<b>{}</b> has been added to bluetext cleaner ignore list.".format(
-                args[0])
+            reply = f"<b>{args[0]}</b> has been added to bluetext cleaner ignore list."
         else:
             reply = "Command is already ignored."
         message.reply_text(reply, parse_mode=ParseMode.HTML)
@@ -122,8 +113,7 @@ def remove_bluetext_ignore(update: Update, context: CallbackContext):
         val = args[0].lower()
         removed = sql.chat_unignore_command(chat.id, val)
         if removed:
-            reply = "<b>{}</b> has been removed from bluetext cleaner ignore list.".format(
-                args[0])
+            reply = f"<b>{args[0]}</b> has been removed from bluetext cleaner ignore list."
         else:
             reply = "Command isn't ignored currently."
         message.reply_text(reply, parse_mode=ParseMode.HTML)
@@ -142,8 +132,8 @@ def add_bluetext_ignore_global(update: Update, context: CallbackContext):
         val = args[0].lower()
         added = sql.global_ignore_command(val)
         if added:
-            reply = "<b>{}</b> has been added to global bluetext cleaner ignore list.".format(
-                args[0])
+            reply = f"<b>{args[0]}</b> has been added to global bluetext cleaner ignore list."
+
         else:
             reply = "Command is already ignored."
         message.reply_text(reply, parse_mode=ParseMode.HTML)
@@ -162,8 +152,8 @@ def remove_bluetext_ignore_global(update: Update, context: CallbackContext):
         val = args[0].lower()
         removed = sql.global_unignore_command(val)
         if removed:
-            reply = "<b>{}</b> has been removed from global bluetext cleaner ignore list.".format(
-                args[0])
+            reply = f"<b>{args[0]}</b> has been removed from global bluetext cleaner ignore list."
+
         else:
             reply = "Command isn't ignored currently."
         message.reply_text(reply, parse_mode=ParseMode.HTML)
